requirements.txt
# Kalshi Prediction Market Sentiment Analyzer - Python 3.10+
kalshi-python>=1.0.0
cryptography>=41.0.0
praw>=7.7.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
openai>=1.0.0
python-dotenv>=1.0.0
